+++ src/xqt-kotlinx-lsp/commonMain/xqt/kotlinx/lsp/base/CancelRequest.kt
@@ -5,6 +5,7 @@ import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.buildJsonObject
 import xqt.kotlinx.rpc.json.protocol.method
+import xqt.kotlinx.rpc.json.protocol.sendNotification
 import xqt.kotlinx.rpc.json.serialization.JsonSerialization
 import xqt.kotlinx.rpc.json.serialization.get
 import xqt.kotlinx.rpc.json.serialization.put
@@ -52,4 +53,41 @@ fun DollarNotification.cancelRequest(
     method = DollarNotification.CANCEL_REQUEST,
     handler = handler,
     paramsSerializer = CancelParams
+)
+
+/**
+ * Cancel an active request.
+ *
+ * A request that got canceled still needs to return from the server and send a response
+ * back. It can not be left open/hanging. This is in line with the JSON RPC protocol that
+ * requires that every request sends a response back. In addition, it allows for returning
+ * partial results on cancel.
+ *
+ * @param params the notification parameters
+ *
+ * @since 2.0.0
+ */
+fun DollarJsonRpcServer.cancelRequest(
+    params: CancelParams
+): Unit = server.sendNotification(
+    method = DollarNotification.CANCEL_REQUEST,
+    params = CancelParams.serializeToJson(params)
+)
+
+/**
+ * Cancel an active request.
+ *
+ * A request that got canceled still needs to return from the server and send a response
+ * back. It can not be left open/hanging. This is in line with the JSON RPC protocol that
+ * requires that every request sends a response back. In addition, it allows for returning
+ * partial results on cancel.
+ *
+ * @param id the request id to cancel
+ *
+ * @since 2.0.0
+ */
+fun DollarJsonRpcServer.cancelRequest(
+    id: JsonIntOrString
+): Unit = cancelRequest(
+    params = CancelParams(id = id)
 )